update plugin name from otel-telemetry-source to otlp
