+++ data-prepper-plugins/otlp-source/data/certificate/test_cert.crt
