+++ data-prepper-plugins/otlp-source/data/certificate/test_decrypted_key.key
