+++ data-prepper-plugins/otlp-source/data/certificate/test_encrypted_key.key
