+++ settings.gradle
@@ -200,3 +200,5 @@ include 'data-prepper-plugins:saas-source-plugins:atlassian-commons'
 include 'data-prepper-plugins:saas-source-plugins:crowdstrike-source'
 include 'data-prepper-plugins:saas-source-plugins:microsoft-office365-source'
 include 'data-prepper-plugins:otlp-sink'
+include 'data-prepper-plugins:otlp-source'
+
